backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
django-channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
whitenoise==6.6.0
gunicorn==21.2.0
